refactor: upgrade torus to use double

+++ src/main/java/com/BudgiePanic/rendering/util/shape/Torus.java
@@ -28,12 +28,12 @@ public class Torus extends BaseShape {
     /**
      * The distance from the toris origin to the torus defining circle.
      */
-    protected final float radius;
+    protected final double radius;
 
     /**
      * The distance from the torus defining circle to the torus surface.
      */
-    protected final float thickness;
+    protected final double thickness;
 
     /**
      * Cached bounding box.
@@ -51,7 +51,7 @@ public class Torus extends BaseShape {
      * @param thickness
      *   The distance from the torus inner circle to the torus surface.
      */
-    public Torus(Matrix4 transform, Material material, float radius, float thickness) { 
+    public Torus(Matrix4 transform, Material material, double radius, double thickness) { 
         super(transform, material); this.radius = radius; this.thickness = thickness; 
         if (thickness >= radius) {
             System.out.println("WARN: torus has thickness is larger than its radius, it won't have a hole");
